Update README.md
This is a project made for the purpose of checking real time availability of Covid-19 vaccines. It uses an API to fetch the data using the date and pin code of your choice. It also answers some frequently asked questions by users. Other components like downloading vaccination certificate are yet to be developed.